requirements.txt
# Runtime dependencies
pyyaml>=6.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.0
# pytest-cov>=4.0
# black>=22.0
# flake8>=5.0
# mypy>=1.0
# build>=0.8
# twine>=4.0
